+++ run_classifier_for_single_snp.m
@@ -2,7 +2,7 @@ function [classifier, success_rate] = run_classifier_for_single_snp(samples, lab
 % RUN_CLASSIFIER_FOR_SINGLE_SNP
 
 if strcmp(cls_method, 'svm-one-vs-one')
-    [classifier, success_rate] = train_svm_one_vs_one(samples, labels);
+    [classifier, success_rate] = train_svm_one_vs_one(double(samples), double(labels));
 end
 end
